// src/openapi/model/api-contractor-project.ts
/* tslint:disable */
/* eslint-disable */

/**
 * mtg_server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 *
 * @export
 * @interface ApiContractorProject
 */
export interface ApiContractorProject {
    /**
     *
     * @type {string}
     * @memberof ApiContractorProject
     */
    recruitingStartDate: string;
    /**
     *
     * @type {string}
     * @memberof ApiContractorProject
     */
    recruitingEndDate: string;
    /**
     *
     * @type {string}
     * @memberof ApiContractorProject
     */
    projectName: string;
    /**
     *
     * @type {number}
     * @memberof ApiContractorProject
     */
    remainingStock: number;
    /**
     *
     * @type {number}
     * @memberof ApiContractorProject
     */
    status: number;
    /**
     *
     * @type {string}
     * @memberof ApiContractorProject
     */
    orderImageUrl: string;
    /**
     *
     * @type {string}
     * @memberof ApiContractorProject
     */
    orderPublicImageUrl: string;
    /**
     *
     * @type {string}
     * @memberof ApiContractorProject
     */
    ordererName: string;
    /**
     *
     * @type {string}
     * @memberof ApiContractorProject
     */
    workStartDate: string;
    /**
     *
     * @type {string}
     * @memberof ApiContractorProject
     */
    restraintStartDate: string;
    /**
     *
     * @type {string}
     * @memberof ApiContractorProject
     */
    restraintEndDate: string;
    /**
     *
     * @type {string}
     * @memberof ApiContractorProject
     */
    areaName: string;
    /**
     *
     * @type {number}
     * @memberof ApiContractorProject
     */
    definitePrice: number;
    /**
     *
     * @type {number}
     * @memberof ApiContractorProject
     */
    projectApplicationId?: number;
    /**
     *
     * @type {number}
     * @memberof ApiContractorProject
     */
    minimumExperienceYears?: number;

    displayPrice?: number;
}