// src/openapi/model/api-contractor-project-change-price.ts
/* tslint:disable */
/* eslint-disable */

/**
 * mtg_server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 *
 * @export
 * @interface ApiContractorProjectChangePrice
 */
export interface ApiContractorProjectChangePrice {
    /**
     *
     * @type {number}
     * @memberof ApiContractorProjectChangePrice
     */
    projectApplicationId: number;
    /**
     *
     * @type {number}
     * @memberof ApiContractorProjectChangePrice
     */
    contractorPrice: number;
}