// src/openapi/model/api-contractor-project-show.ts
/* tslint:disable */
/* eslint-disable */
/**
 * mtg_server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import {ApiProject} from './api-project';
import {ApiProjectApplication} from './api-project-application';
import {ApiProjectFile} from './api-project-file';
import {ApiUser} from './api-user';

/**
 *
 * @export
 * @interface ApiContractorProjectShow
 */
export interface ApiContractorProjectShow {
    /**
     *
     * @type {string}
     * @memberof ApiContractorProjectShow
     */
    areaName?: string;
    /**
     *
     * @type {number}
     * @memberof ApiContractorProjectShow
     */
    projectId?: number;
    /**
     *
     * @type {number}
     * @memberof ApiContractorProjectShow
     */
    projectApplicationId?: number;
    /**
     *
     * @type {number}
     * @memberof ApiContractorProjectShow
     */
    ordererId?: number;
    /**
     *
     * @type {Array<string>}
     * @memberof ApiContractorProjectShow
     */
    genreNames?: Array<string>;
    /**
     *
     * @type {number}
     * @memberof ApiContractorProjectShow
     */
    projectStock?: number;
    /**
     *
     * @type {ApiUser}
     * @memberof ApiContractorProjectShow
     */
    orderer?: ApiUser;
    /**
     *
     * @type {ApiProject}
     * @memberof ApiContractorProjectShow
     */
    project?: ApiProject;
    /**
     *
     * @type {ApiProjectApplication}
     * @memberof ApiContractorProjectShow
     */
    projectApplication?: ApiProjectApplication;
    /**
     *
     * @type {Array<ApiProjectFile>}
     * @memberof ApiContractorProjectShow
     */
    projectFile?: Array<ApiProjectFile>;

    displayPrice?: number;
}