// src/openapi/model/api-project-search.ts
/* tslint:disable */
/* eslint-disable */
/**
 * mtg_server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 *
 * @export
 * @interface ApiProjectSearch
 */
export interface ApiProjectSearch {
    /**
     *
     * @type {number}
     * @memberof ApiProjectSearch
     */
    id: number;
    /**
     *
     * @type {string}
     * @memberof ApiProjectSearch
     */
    createdAt: string;
    /**
     * 
     * @type {string}
     * @memberof ApiProjectSearch
     */
    updatedAt: string;
    /**
     *
     * @type {string}
     * @memberof ApiProjectSearch
     */
    recruitingStartDate: string;
    /**
     *
     * @type {string}
     * @memberof ApiProjectSearch
     */
    recruitingEndDate: string;
    /**
     *
     * @type {string}
     * @memberof ApiProjectSearch
     */
    projectName: string;
    /**
     *
     * @type {string}
     * @memberof ApiProjectSearch
     */
    userProfileImageUrl: string;
    /**
     *
     * @type {string}
     * @memberof ApiProjectSearch
     */
    userProfilePublicImageUrl: string;
    /**
     *
     * @type {string}
     * @memberof ApiProjectSearch
     */
    workStartDate: string;
    /**
     *
     * @type {string}
     * @memberof ApiProjectSearch
     */
    areaName: string;
    /**
     *
     * @type {number}
     * @memberof ApiProjectSearch
     */
    minPrice: number;
    /**
     *
     * @type {number}
     * @memberof ApiProjectSearch
     */
    maxPrice: number;
    /**
     *
     * @type {number}
     * @memberof ApiProjectSearch
     */
    matchingPoint: number;
    /**
     *
     * @type {boolean}
     * @memberof ApiProjectSearch
     */
    activeHeart: boolean;
}