// src/openapi/model/api-user-favorite-project-to-add.ts
/* tslint:disable */
/* eslint-disable */
/**
 * mtg_server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 *
 * @export
 * @interface ApiUserFavoriteProjectToAdd
 */
export interface ApiUserFavoriteProjectToAdd {
    /**
     *
     * @type {number}
     * @memberof ApiUserFavoriteProjectToAdd
     */
    userId?: number;
    /**
     *
     * @type {number}
     * @memberof ApiUserFavoriteProjectToAdd
     */
    projectId: number;
}